requirements.txt
asgiref==3.4.1
autopep8==1.5.7
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==8.0.1
cryptography==3.4.8
Deprecated==1.2.13
fastapi==0.68.1
gcloud==0.18.3
googleapis-common-protos==1.53.0
h11==0.12.0
httplib2==0.19.1
idna==3.2
jwcrypto==1.0
oauth2client==4.1.3
protobuf==3.18.0rc2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
pycryptodome==3.10.1
pydantic==1.8.2
pyparsing==2.4.7
Pyrebase4==4.5.0
python-jwt==3.3.0
requests==2.26.0
requests-toolbelt==0.9.1
rsa==4.7.2
six==1.16.0
starlette==0.14.2
toml==0.10.2
typing-extensions==3.10.0.2
urllib3==1.26.6
uvicorn==0.15.0
wrapt==1.12.1
